Rename file exporting Styles from styleTypes to Styles
